UI/UX developed:

+++ workoutapp/api/routes/workout.js
@@ -8,6 +8,64 @@ const verifyToken = require('../middleware/verify-token');
 
 
 
+//Getting Workouts For user
+router.get('/', async(req, res) => {
+    try {
+
+        //Getting all workouts for user
+        const user = await User.findByUsername("Matt").populate({
+            path: 'workouts',
+            populate: {
+                path: 'exercises',
+                populate: 'sets'
+            }
+        }).exec();
+        
+        res.json({
+            success: true,
+            workouts: user.workouts
+        })
+
+        
+    } catch (err) {
+        res.status(500).json({
+            success: false,
+            message: err.message
+        })
+    }
+
+
+
+})
+
+//Getting Workouts For user
+router.get('/:id', async(req, res) => {
+    try {
+
+        //Getting all workouts for user
+        const workout = await Workout.findById(req.params.id);
+
+            res.json({
+                success: true,
+                workout: workout
+            })
+
+        
+        
+
+        
+    } catch (err) {
+        res.status(500).json({
+            success: false,
+            message: err.message
+        })
+    }
+
+
+
+})
+
+
 //Creating Workout 
 router.post('/create', async (req, res) => {
     const user = await User.findByUsername("Matt");
@@ -17,8 +75,10 @@ router.post('/create', async (req, res) => {
         name: req.body.name,
         exercises: [],
         userId: user._id,
-        isCompleted: false
+        isCompleted: false,
+        dateScheduled: req.body.dateScheduled
     });
+
     await workout.save()
     await user.workouts.push(workout);
     await user.save()
@@ -36,26 +96,37 @@ router.delete('/', async(req, res) => {
     try {
         const user = await User.findByUsername("Matt");
 
-    
-        let userWorkouts = user.workouts
-        var filtered = userWorkouts.filter(function(value, index, arr){ 
-            return value._id != req.body.workoutId;
-        });
+        if(user.workouts !== null) {
+            let userWorkouts = user.workouts
+                var filtered = userWorkouts.filter(function(value, index, arr){ 
+                    return value._id != req.body.workoutId;
+            });            
+        }
+ 
 
         const workout = await Workout.findOne({_id: req.body.workoutId})
 
-        //Deleting exercises inside workout
-        for(let exerciseId of workout.exercises)
-        {
-            const exercise = await Exercise.findById(exerciseId);
+        console.log(workout.exercises)
 
-            //Deleting Sets within Exercises
-            for(let setId of exercise.sets)
+        //Deleting exercises inside workout
+        if(workout.exercises) {
+
+            for(let exerciseId of workout.exercises)
             {
-                await Set.findOneAndDelete({_id: setId});
-            }
-            await Exercise.findOneAndDelete({_id: exerciseId});
+                const exercise = await Exercise.findById(exerciseId);
+
+                //Deleting Sets within Exercises
+                if(exercise.sets !== null) {
+                    for(let setId of exercise.sets)
+                    {
+                        await Set.findOneAndDelete({_id: setId});
+                    }                    
+                }
+
+                await Exercise.findOneAndDelete({_id: exerciseId});
+            }            
         }
+
 
         await Workout.findOneAndDelete({_id: req.body.workoutId});
 
@@ -96,10 +167,29 @@ router.put('/', async (req, res) => {
             message: err.message
         });
     }
-
-
-
 });
+
+
+//Changing isCompleted to true or false
+router.put('/isCompleted', async (req, res) => {
+    try {
+        const exercise = await Workout.findByIdAndUpdate(req.body.workoutId, {
+            isCompleted: req.body.isCompleted
+        })
+
+        await exercise.save();
+
+        res.json({
+            success: true,
+            message: "You have updated is Completed for exercise"
+        })
+    } catch (err) {
+        res.status(500).json({
+            success: false,
+            message: err.message
+        })
+    }
+})
